+++ Cargo.toml
@@ -37,7 +37,7 @@ members = [
     "packages/fullstack/examples/axum-router",
     "packages/fullstack/examples/axum-desktop",
     "packages/fullstack/examples/axum-auth",
-    "packages/fullstack/examples/cloudflare-workers-hello-world",
+    "packages/fullstack/examples/cloudflare-hello-world",
     "packages/fullstack/examples/salvo-hello-world",
     "packages/fullstack/examples/warp-hello-world",
     "packages/fullstack/examples/static-hydrated",
