+++ packages/fullstack/Cargo.toml
@@ -84,7 +84,7 @@ mobile = ["dioxus-mobile"]
 warp = ["dep:warp", "server"]
 axum = ["dep:axum", "tower-http", "server"]
 salvo = ["dep:salvo", "server", "http-body-util"]
-cloudflare-workers = ["worker", "server"]
+cloudflare = ["worker", "server"]
 server = ["server_fn/ssr", "dioxus_server_macro/server", "tokio", "tokio-util", "tokio-stream", "dioxus-ssr", "dioxus-ssr/incremental", "tower", "hyper", "http", "tower-layer", "anymap", "tracing-futures", "pin-project", "thiserror", "dioxus-cli-config"]
 default-tls = ["server_fn/default-tls"]
 rustls = ["server_fn/rustls"]
