+++ packages/fullstack/examples/cloudflare-hello-world/package.json
