+++ packages/fullstack/examples/cloudflare-hello-world/src/lib.rs
