+++ packages/fullstack/examples/cloudflare-hello-world/wrangler.toml
