+++ packages/fullstack/src/adapters/cloudflare_adapter.rs
