+++ packages/fullstack/src/adapters/mod.rs
@@ -14,10 +14,10 @@
 pub mod axum_adapter;
 // #[cfg(feature = "salvo")]
 // pub mod salvo_adapter;
+#[cfg(feature = "cloudflare")]
+pub mod cloudflare_adapter;
 #[cfg(feature = "warp")]
 pub mod warp_adapter;
-#[cfg(feature = "cloudflare-workers")]
-pub mod worker_adapter;
 
 use http::StatusCode;
 use server_fn::{Encoding, Payload};
