+++ packages/fullstack/src/lib.rs
@@ -53,15 +53,12 @@ pub mod prelude {
     // #[cfg(feature = "salvo")]
     // #[cfg_attr(any(docsrs, feature = "nightly-doc"), doc(cfg(feature = "salvo")))]
     // pub use crate::adapters::salvo_adapter::*;
+    #[cfg(feature = "cloudflare")]
+    #[cfg_attr(any(docsrs, feature = "nightly-doc"), doc(cfg(feature = "cloudflare")))]
+    pub use crate::adapters::cloudflare_adapter::*;
     #[cfg(feature = "warp")]
     #[cfg_attr(any(docsrs, feature = "nightly-doc"), doc(cfg(feature = "warp")))]
     pub use crate::adapters::warp_adapter::*;
-    #[cfg(feature = "cloudflare-workers")]
-    #[cfg_attr(
-        any(docsrs, feature = "nightly-doc"),
-        doc(cfg(feature = "cloudflare-workers"))
-    )]
-    pub use crate::adapters::worker_adapter::*;
     use crate::hooks;
     #[cfg(not(feature = "server"))]
     #[cfg_attr(
